fixes issue #14 compiling on Win 7 now works
This commit adds custom build.bat for Windows and changes the code
so that it can be compiled on Windows.

The resulting exe files run without issues in Powershell, cmd.exe
and Cygwin.

The code is not the cleanest and could be improved and there are
also a few Notes and Todos throughout it that could be addressed
in the future.

For now the big breakthrough is that "it compiles" on Windows.